Added simple function for http request in common directory
