+++ gtkm/common/__init__.py
@@ -1,1 +1,2 @@
 from .routing import gen_url
+from .http_request import get_endpoint_data
